replace zustand state by local storage hook
