+++ package.json
@@ -10,6 +10,7 @@
     "@wireapp/store-engine-sqleet": "1.8.9",
     "@wireapp/webapp-events": "0.15.2",
     "amplify": "https://github.com/wireapp/amplify#head=master",
+    "beautiful-react-hooks": "^3.11.2",
     "classnames": "2.3.2",
     "copy-webpack-plugin": "11.0.0",
     "core-js": "3.25.5",
