+++ src/script/page/MainContent/panels/preferences/OptionPreferences.tsx
@@ -32,8 +32,8 @@ import {t} from 'Util/LocalizerUtil';
 import {PreferencesPage} from './components/PreferencesPage';
 import {PreferencesRadio} from './components/PreferencesRadio';
 import {PreferencesSection} from './components/PreferencesSection';
-import {RootFontSize, useFontSizeState} from './useFontSizePreferenceState';
 
+import {RootFontSize, useRootFontSize} from '../../../../hooks/useRootFontSize';
 import {PropertiesRepository} from '../../../../properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from '../../../../properties/PropertiesType';
 import {UserState} from '../../../../user/UserState';
@@ -68,7 +68,7 @@ const OptionPreferences: React.FC<OptionPreferencesProps> = ({
   const [optionDarkMode, setOptionDarkMode] = useState<boolean>(settings.interface.theme === ThemeViewModelThemes.DARK);
   const [optionSendPreviews, setOptionSendPreviews] = useState<boolean>(settings.previews.send);
   const [optionNotifications, setOptionNotifications] = useState<NotificationPreference>(settings.notifications);
-  const {currentRootFontSize, setCurrentRootFontSize} = useFontSizeState(state => state.rootFontSize);
+  const [currentRootFontSize, setCurrentRootFontSize] = useRootFontSize();
 
   useEffect(() => {
     const updateProperties = ({settings}: WebappProperties): void => {
